+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Add `ImagesRemoveOptsBuilder::ignore`
 * Add `ImagesRemoveOptsBuilder::lookup_manifest`
 * Add `ImagePushOptsBuilder::quiet`
+* Add `NetworkCreateOptsBuilder::created`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync